requirements-dev.txt
# =============================================================================
# REQUIREMENTS DE DÉVELOPPEMENT - Observations Nids
# =============================================================================
# Installation : pip install -r requirements.txt -r requirements-dev.txt
# =============================================================================

# -----------------------------------------------------------------------------
# Outils de formatage et linting
# -----------------------------------------------------------------------------
black==25.1.0
ruff==0.12.12

# -----------------------------------------------------------------------------
# Analyse de types
# -----------------------------------------------------------------------------
mypy==1.17.1
mypy-extensions==1.0.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
types-pytz==2025.2.0.20250809
types-PyYAML==6.0.12.20241230
typing-inspection==0.4.0

# -----------------------------------------------------------------------------
# Outils de débogage Django
# -----------------------------------------------------------------------------
django-debug-toolbar==5.1.0
django-extensions==3.2.3

# -----------------------------------------------------------------------------
# Tests
# -----------------------------------------------------------------------------
pytest==8.3.5
pytest-django==4.9.0
pytest-cov==6.0.0

# -----------------------------------------------------------------------------
# Analyse de code et graphiques de dépendances
# -----------------------------------------------------------------------------
pyan3==1.2.0
pygraphviz==1.14

# -----------------------------------------------------------------------------
# Gestion des dépendances
# -----------------------------------------------------------------------------
pip-tools==7.4.1

# -----------------------------------------------------------------------------
# Analyse de données (pour analyse de logs ou datasets)
# -----------------------------------------------------------------------------
pandas==2.3.2
pandas-stubs==2.3.2.250827
numpy==2.3.3

# -----------------------------------------------------------------------------
# Build tools
# -----------------------------------------------------------------------------
build==1.2.2.post1
pyproject_hooks==1.2.0
platformdirs==4.4.0
pathspec==0.12.1
setuptools==76.0.0
wheel==0.45.1
